Add the common blocks to Sager. Added Avnet
There is a lot of duplication here, but I prefer that for now, so that any special requests we get for a particular site area easy to achieve without messing up any other sites.

The only downside is that we may have to apply “fixes” or updates to multile sites, but at 4 sites, this is not particulary difficult.

Add Avnet to installed apps, create migration.

Add the common blocks to Sager

We need this setup so that the demo team can start creating pages right away.