Revert the "javascript" based scroll logic. According to our UI/UX designer, Bootstrap should have a built-in mechanism to achieve the effect we are looking for, so I don't want to use javascript and magic values.
Refactored the main html to support sticky menu
This fixes issues with the sticky menu, and requires no custom JS for that. It uses BS4 only.
I've also updated the HTML to fix issues with the menu shown for mobile. I had to re-peat the html, but I'm not too worried about that.
Update html and css for the mobile jump to section

- Update to the javascript to make the open/close work
- Cleaned up the html so that both desktop and mobile versions of the menu work
- Cleaned up the scss so we have a mobile version and non-mobile version

Remove unused jump section from desktop sidebar html