Refactor template blocks for rendering modules. Update cache key for hierarchy data, and re-name data var
I refactored the blocks for rendering the modules so that it is easier to render some react component in the common datapages template, but then easily add additonal ones later.

Modified the key for the hierarchy cache

Namedspaced the variable used for hierarchy data. It was just ‘data’ which is not that great. added a ‘datapages’ namespace, and put the variable there named ‘hierarchy_data’
It won’t conflice now.